ibus: Request cursor rectangle whenever the preedit text is changed
Cursor rectangle should be updated before a preedit candidate window is
shown. For applications using custom widgets with an inputMethodQuery
callback, ibus preedit candidate window is not placed correctly unless
the application calls QInputMethod::update() whenever a possible change
for the cursor rectangle occurs. However, as the upadte call involves
DBus message passing in the case of ibus, it brings severe performance
drawbacks. A better approach is that the Qt platform plugin requests the
cursor rectangle when necessary. For ibus, the preedit cursor position
is changed before a candidate window is shown, so issue a request
whenever the preedit text is changed can fix the misplaced candidate
window.

Change-Id: I1fdb3820c2085a2997c1ea735b1fb1a0940788f8